feat: CLIN-1763 autocomplete et traduction
