+++ src/locales/en.json
@@ -9,6 +9,10 @@
   "affected": "Affected",
   "allelic.ratio": "Allelic Ratio",
   "annotationUpdate": "Annotation date",
+  "assignment.dropdown.actions.clear": "No assignment",
+  "assignment.dropdown.select.searchPlaceholder": "Search",
+  "assignment.tooltip": "Assign this prescription",
+  "assignment.tooltip.disable": "No assignment",
   "birthdate": "Birth Date",
   "cancel": "Cancel",
   "cant.have.two.file.number.same.patient": "Cannot have two different file numbers for the same patient",
