+++ src/locales/fr.json
@@ -8,6 +8,10 @@
   "acmg.criteria": "Critères ACMG",
   "affected": "Atteint",
   "allelic.ratio": "Ratio allélique",
+  "assignment.dropdown.actions.clear": "Aucune assignation",
+  "assignment.dropdown.select.searchPlaceholder": "Rechercher",
+  "assignment.tooltip": "Assigner cette prescription",
+  "assignment.tooltip.disable": "Aucune assignation",
   "annotationUpdate": "Date d’annotation",
   "birthdate": "Date de naissance",
   "cancel": "Annuler",
