+++ src/utils/translation.ts
@@ -1,4 +1,5 @@
 import intl from 'react-intl-universal';
+import { IDictionary } from '@ferlab/ui/core/components/Assignments/types';
 import { IDictionary as FiltersDict } from '@ferlab/ui/core/components/filters/types';
 import { IProTableDictionary } from '@ferlab/ui/core/components/ProTable/types';
 import { IDictionary as QueryBuilderDict } from '@ferlab/ui/core/components/QueryBuilder/types';
@@ -71,6 +72,15 @@ export const getProTableDictionary = (): IProTableDictionary => ({
     previous: intl.get('previous'),
     next: intl.get('next'),
     view: '',
+  },
+});
+
+export const getAssignmentDictionary = (): IDictionary => ({
+  actions: {
+    clear: intl.get('assignment.dropdown.actions.clear'),
+  },
+  select: {
+    searchPlaceholder: intl.get('assignment.dropdown.select.searchPlaceholder'),
   },
 });
