+++ src/views/Prescriptions/utils/export.ts
@@ -1,5 +1,8 @@
+import { TPractitionnerInfo } from '@ferlab/ui/core/components/Assignments/types';
 import { ISyntheticSqon } from '@ferlab/ui/core/data/sqon/types';
+import { Practitioner, PractitionerRole } from 'api/fhir/models';
 import { findDonorById } from 'graphql/variants/selector';
+import { filter, find } from 'lodash';
 import get from 'lodash/get';
 import { renderCNVToString } from 'views/Cnv/Exploration/variantColumns';
 import { renderToString as renderConsequencesToString } from 'views/Snv/components/ConsequencesCell/index';
@@ -232,4 +235,29 @@ export const makeFilenameDatePart = (date = new Date()) => {
     date.getUTCSeconds(),
   ]);
   return `${prefixes}T${suffixes}Z`;
+};
+
+export const getEmailfromPractionnerRole = (practitionerRole: PractitionerRole) =>
+  practitionerRole?.telecom.find((t) => t.system === 'email');
+
+export const getPractitionerInfoList = (practitionerRolesBundle?: any[]): TPractitionnerInfo[] => {
+  const practitionerRoles: PractitionerRole[] = filter(
+    practitionerRolesBundle,
+    (p) => p.resourceType === 'PractitionerRole',
+  );
+  return practitionerRoles.reduce((acc: TPractitionnerInfo[], curr: PractitionerRole) => {
+    const practitioner: Practitioner = find(
+      practitionerRolesBundle,
+      (p) => p.id === curr.practitioner.reference.split('/')[1],
+    );
+    const email = getEmailfromPractionnerRole(curr);
+    const obj: TPractitionnerInfo = {
+      practitionerRoles_Id: curr.id,
+      name: practitioner.name,
+      email: email?.value,
+      ldm: curr.organization.reference.split('/')[1],
+    };
+
+    return [...acc, obj];
+  }, []);
 };
